feat: adds fallback procedure when data connection changes or is lost (#73)
When a user walks outside of a wifi, the application used to lose
connectivity with the agent... without letting the Explorer know
anything about the issue. This PR:

lets the user know when a connection is lost (not switching here, but
simply lost)
without adding a reconnect button, this PR automatically attempts to
reconnect as soon as a connection change happens while keeping the user
aware of that pending reconnection.
uses a much more reliable reconnection process than a simple "legacy"
reconnect. (we use a WebRTC restartIce).
adds a mechanism to lets the user know when a reconnection failed (This
information wasn't getting back to the Explorer before)... although we
still have to improve that "legacy" reconnection process which not only
fails often, but doesn't provide much feedback.