fix(PubNub): fixed pubnub reconnection issue (#74)
Once we have a disconnection due to network glitch or network change,
PubNub wasn't attempting to reconnect. This PR fixes this.

> Based on
https://www.pubnub.com/docs/sdks/dart/api-reference/publish-and-subscribe#description-2,
it sounds like if the PubNub client gets disconnected from a channel for
whatever reason, it won’t automatically reconnect. If we want it to do
that, we need to specify a RetryPolicy when creating the client, similar
to https://github.com/pubnub/dart/issues/19#issuecomment-692153308.