fix(platform_client & room): use of a common location throttling mechanism to avoid concurrent updates (#85)
We have two APIs to update Explorer's locations: 

1. `Room.updateLocation` which uses a rabbitMQ API to communicate the
update. This function is used to update location during a call.
2. `PlatformClient.inquireForGPSActivatedOffer` which uses a REST
endpoint to know if there is a new GPS activated offer at a specific
location. This function is used before calls and in the background.

These CPU intensive calls execute the same code on the backend which can
cause a "endpoint modified concurrently" PlatformLocalizedException if a
user ends up to execute this code while his previous call is not
finished.

This PR fixes this issue by using a common throttling mechanism between
these two functions and by increasing the delay between the possible
calls from 1 second to 1.5 seconds.