fix(room): fixed the source of the video pixelation when starting or stoping a presentation (#87)
As reported in some of the following links:

https://airasproductt-hqa4190.slack.com/archives/C044B4WAK1U/p1689980551973879?thread_ts=1689976009.201869&cid=C044B4WAK1U

https://airasproductt-hqa4190.slack.com/archives/C044B4WAK1U/p1690404970446239

https://airasproductt-hqa4190.slack.com/archives/C044B4WAK1U/p1690417428049009

The problem:
Starting or stoping a presentation could cause a pixelation of the video
on the Agent side.

Observations:
After some experimentation, I discovered that, once I succeed to start a
presentation which was causing the pixelation, the video dimensions
failed to change although it should have. This "resolution" change seems
to be at the source of the problem.
We also noted that a reconnect always seems to fix the issue.

The fix:
After attempting to address the "resolution" problem without success, I
decided to change the way we start and end presentations for a track
re-creation (which is the equivalent to a partial reconnect).

Side effects:
The process to start and stop a presentation is longer and causes the
Agent to lose the audio and video for about 2 seconds.
There is no longer any kind of pixelation.