fix: fixed issue where calls could start without video (#106)
# Description

This PR addresses the issue described
[here](https://aira.slack.com/archives/C01LB57N74Z/p1706390196107819).

Here is the flow of events as a call start in a very simple form:

1. Client: Room.createServiceRequest is called.
2. Room: The service status request is both received from a RabbitMQ
message and polled
3. Room: We get the Request ServiceRequest status and do nothing
4. Room: We get the Assigned ServiceRequest status and notify listeners.
Step 5 & 6 happen concurrently.
5. Client: We create the video track and join the room (which is
creating the video transceiver)
6. Room: Continues the connection process.
7. Room: We get the Started ServiceRequest status and update the
ParticipantStatus and notify listeners.

Things to keep in mind as you continue this reading:

A) `Room.setVideoMute` does call `Room._updateParticipantStatus` but the
later only executes if the ServiceRequest status is set to `Started`

B) `Room._updateParticipantStatus` will set the Participant status to
`PRIVACY_AUDIO_ONLY` when `_isVideoMuted` is true or when
`_hasVideoTrack` is false. In our case, we expect this is the later
which causes the VI to see the video as muted as our Explorers say they
did not mute the video.

## Possibility no1:
We update the VideoMute status before we have the video which leads to a
PRIVACY_AUDIO_ONLY participant status, which in turn causes the VI to
see the "video muted" icon instead of the video.

It is unclear to me when the call to `Room.createServiceRequest` returns
and this could vary on device types, concurrency condition, etc. I
mention this because this is just when that function returns that we
used to call `Room.setVideoMute` to set the starting mute status in
Room. The question arise because of item A). So if the video was never
muted on the client, we would need these condition to be true to cause
the problem:

1. We don't have a video track yet.
2. The ServiceRequest Status is: Started

These conditions can only happen if the client has finished to create
the video track and the ServiceRequest is now set to True.

## Possibility no2:
Could it be possible that the MediaStream gets created without video?
Probably not as we wait for the call to `getMediaStream` (and all its
wrapping functions) to complete, but nothing warranties that the video
track is within the video tracks at that time. This would affect both
the video mute status (see item B) and our ability to create an outgoing
video transceiver (this is done as we join the room)

## Possibility no3:
Could it be possible that we join the room after the ServiceRequest
status is set to "Started"? This could be possible if the time to create
the video track is longer than the time for Agent Dashboard to be ready
to receive. I didn't go through the whole code in Platform to figure out
this possibility, but we do have this comment in
`ServiceRequestManager`: "// START is called once dashboard is loaded
and ready to receive"
This would mean that the `_localStream` would be null when the
ServiceRequest status is "Started".

## Proposed solution:
Lets address the 3 possibilities. Only the 2 last one can be resolved in
Flutter_aira though.