fix: addressing the 'Unable to RTCPeerConnection::addCandidate: Error The remote description was null' issue (#121)
WebRTC depends on a series of events. The SDP offer is a step which must
happen before we start registering ICE Candidates (routes). This fix
basically stores potential ICE Candidates until we have a SDP offer
accepted. Then it registers the stored Candidates and allow subsequents
offer to be added directly.


https://linear.app/airaio/issue/EXPO-346/sentry-the-remote-description-was-null